fix: 拆分 fixF2 方法
