+++ @types/f2-tool.d.ts
@@ -1,0 +1,1 @@
+export function fixF2(F2:any): any
