+++ @types/index.d.ts
@@ -1,1 +1,2 @@
 export { default as F2Canvas } from './f2-canvas'
+export { fixF2 } from './f2-tool'
