+++ package.json
@@ -43,22 +43,23 @@
 	"author": "xioxin",
 	"license": "MIT",
 	"devDependencies": {
-		"@babel/core": "^7.1.2",
+		"@antv/f2": "^3.3.1",
+		"@babel/core": "7.1.2",
 		"@babel/plugin-proposal-class-properties": "^7.1.0",
 		"@babel/plugin-transform-react-jsx": "^7.0.0",
 		"@babel/preset-env": "^7.1.0",
-		"@tarojs/components": "1.3.0-beta.5",
-		"@tarojs/plugin-babel": "1.3.0-beta.5",
-		"@tarojs/plugin-csso": "1.3.0-beta.5",
-		"@tarojs/plugin-sass": "1.3.0-beta.5",
-		"@tarojs/plugin-uglifyjs": "1.3.0-beta.5",
-		"@tarojs/router": "1.3.0-beta.5",
-		"@tarojs/taro": "1.3.0-beta.5",
-		"@tarojs/taro-alipay": "1.3.0-beta.5",
-		"@tarojs/taro-h5": "1.3.0-beta.5",
-		"@tarojs/taro-swan": "1.3.0-beta.5",
-		"@tarojs/taro-weapp": "1.3.0-beta.5",
-		"@tarojs/webpack-runner": "1.2.27-beta.0",
+		"@tarojs/components": "1.3.22",
+		"@tarojs/plugin-babel": "1.3.22",
+		"@tarojs/plugin-csso": "1.3.22",
+		"@tarojs/plugin-sass": "1.3.22",
+		"@tarojs/plugin-uglifyjs": "1.3.22",
+		"@tarojs/router": "1.3.22",
+		"@tarojs/taro": "1.3.22",
+		"@tarojs/taro-alipay": "1.3.22",
+		"@tarojs/taro-h5": "1.3.22",
+		"@tarojs/taro-swan": "1.3.22",
+		"@tarojs/taro-weapp": "1.3.22",
+		"@tarojs/webpack-runner": "1.3.22",
 		"@types/react": "^16.4.6",
 		"@types/sinon": "^5.0.5",
 		"@types/webpack-env": "^1.13.6",
@@ -69,13 +70,14 @@
 		"babel-plugin-transform-decorators-legacy": "^1.3.4",
 		"babel-plugin-transform-jsx-stylesheet": "^0.6.5",
 		"babel-plugin-transform-object-rest-spread": "^6.26.0",
+		"babel-plugin-transform-remove-strict-mode": "^0.0.2",
 		"babel-preset-env": "^1.6.1",
 		"cross-env": "^5.2.0",
 		"eslint": "^4.19.1",
-		"eslint-config-taro": "1.3.0-beta.5",
+		"eslint-config-taro": "1.3.22",
 		"eslint-plugin-import": "^2.12.0",
 		"eslint-plugin-react": "^7.8.2",
-		"eslint-plugin-taro": "1.3.0-beta.5",
+		"eslint-plugin-taro": "1.3.22",
 		"husky": "^1.1.2",
 		"jest": "^23.6.0",
 		"lint-staged": "^8.0.3",
@@ -85,8 +87,7 @@
 		"regenerator-runtime": "^0.12.1",
 		"simulant": "^0.2.2",
 		"sinon": "^7.1.0",
-		"typescript": "^3.0.1",
-		"@antv/f2": "^3.3.1"
+		"typescript": "^3.0.1"
 	},
 	"dependencies": {
 		"wolfy87-eventemitter": "^5.2.5"
