+++ src/index.ts
@@ -1,1 +1,2 @@
 export { default as F2Canvas } from './components/f2-canvas/f2-canvas'
+export { fixF2 } from './common/f2-tool'
