+++ src/pages/charts/bar/index.tsx
@@ -1,6 +1,7 @@
 import Taro, { Component, Config } from '@tarojs/taro';
 import F2Canvas from "../../../components/f2-canvas/f2-canvas";
 import {View} from '@tarojs/components';
+import {fixF2} from "../../../common/f2-tool"
 
 const F2 = require("@antv/f2");
 
@@ -13,7 +14,7 @@ export default class Index extends Component {
   state = { };
 
     initChart (canvas, width, height) {
-    F2Canvas.fixF2(F2);
+    fixF2(F2);
     const chart = new F2.Chart({
       el: canvas,
       width,
