+++ src/pages/charts/column-select/index.tsx
@@ -1,5 +1,6 @@
 import Taro, { Component, Config } from '@tarojs/taro';
 import F2Canvas from "../../../components/f2-canvas/f2-canvas";
+import {fixF2} from "../../../common/f2-tool"
 import {View} from '@tarojs/components';
 
 const F2 = require("@antv/f2");
@@ -12,7 +13,7 @@ export default class Index extends Component {
 
   state = { };
     initChart (canvas, width, height) {
-    F2Canvas.fixF2(F2);
+    fixF2(F2);
     const data = [
       { year: '1951 年', sales: 38 },
       { year: '1952 年', sales: 52 },
