+++ src/pages/charts/pie-select/index.tsx
@@ -1,5 +1,6 @@
 import Taro, { Component, Config } from '@tarojs/taro';
 import F2Canvas from "../../../components/f2-canvas/f2-canvas";
+import {fixF2} from "../../../common/f2-tool"
 import {View, Text} from '@tarojs/components';
 
 
@@ -17,8 +18,8 @@ export default class Index extends Component {
     message: '',
   };
 
-    initChart (canvas, width, height) {
-    F2Canvas.fixF2(F2);
+  initChart (canvas, width, height) {
+    fixF2(F2);
     var data = [
       { name: '芳华', percent: 0.4, a: '1' },
       { name: '妖猫传', percent: 0.2, a: '1' },
