+++ src/pages/charts/radial-bar/index.tsx
@@ -1,5 +1,6 @@
 import Taro, { Component, Config } from '@tarojs/taro';
 import F2Canvas from "../../../components/f2-canvas/f2-canvas";
+import {fixF2} from "../../../common/f2-tool"
 import {View} from '@tarojs/components';
 
 
@@ -20,8 +21,8 @@ export default class Index extends Component {
   state = { };
 
 
-    initChart (canvas, width, height) {
-    F2Canvas.fixF2(F2);
+  initChart (canvas, width, height) {
+    fixF2(F2);
     // customize shape and animation
     Shape.registerShape('interval', 'tick', {
       draw(cfg, container) {
