+++ src/pages/charts/scroll-line/index.tsx
@@ -2,6 +2,7 @@ import Taro, { Component, Config } from '@tarojs/taro';
 import {View} from '@tarojs/components';
 import data from '../../../data/sroll-line.js'
 import F2Canvas from "../../../components/f2-canvas/f2-canvas";
+import {fixF2} from "../../../common/f2-tool"
 
 const F2 = require("@antv/f2");
 
@@ -19,7 +20,7 @@ export default class Index extends Component {
   state = { };
 
   initChart(canvas, width, height) {
-    F2Canvas.fixF2(F2);
+    fixF2(F2);
     const chart = new F2.Chart({
       el: canvas,
       width,
