+++ src/pages/charts/stackBar/index.tsx
@@ -1,5 +1,6 @@
 import Taro, { Component, Config } from '@tarojs/taro';
 import F2Canvas from "../../../components/f2-canvas/f2-canvas";
+import {fixF2} from "../../../common/f2-tool"
 import {View} from '@tarojs/components';
 
 
@@ -15,8 +16,8 @@ export default class Index extends Component {
 
   state = { };
 
-    initChart (canvas, width, height) {
-    F2Canvas.fixF2(F2);
+  initChart (canvas, width, height) {
+    fixF2(F2);
     const data = [
       { State: 'WY', 年龄段: '小于5岁', 人口数量: 25635 },
       { State: 'WY', 年龄段: '5至13岁', 人口数量: 1890 },
