+++ src/pages/charts/steps-pan/index.tsx
@@ -1,5 +1,6 @@
 import Taro, { Component, Config } from '@tarojs/taro';
 import F2Canvas from "../../../components/f2-canvas/f2-canvas";
+import {fixF2} from "../../../common/f2-tool"
 import data from '../../../data/steps-pan.js'
 import {View} from '@tarojs/components';
 
@@ -23,7 +24,7 @@ export default class Index extends Component {
   state = { };
 
     initChart (canvas, width, height) {
-    F2Canvas.fixF2(F2);
+    fixF2(F2);
     const originDates = [];
     data.forEach(obj => {
       if (obj.date >= '2018-05-01') {
